Update neovim installation for Ubuntu

+++ neovim/neovim-setup.sh
@@ -9,11 +9,16 @@ sudo apt install -y python3-venv
 # Function to install Neovim if it's not installed
 install_nvim() {
   echo "Neovim is not installed. Installing..."
-  curl -LO https://github.com/neovim/neovim/releases/latest/download/nvim-linux64.tar.gz
-  sudo rm -rf /opt/nvim
-  sudo tar -C /opt -xzf nvim-linux64.tar.gz
-  sudo ln -sf /opt/nvim-linux64/bin/nvim /usr/local/bin/nvim  # Make Neovim accessible via the nvim command
-  echo "Neovim installation completed."
+  if [  -n "$(cat /etc/os-release | grep Ubuntu)" ]; then
+    sudo apt-get install neovim
+  else
+    echo "Installing neovim manually -- operating system not supported"
+    curl -LO https://github.com/neovim/neovim/releases/latest/download/nvim-linux64.tar.gz
+    sudo rm -rf /opt/nvim
+    sudo tar -C /opt -xzf nvim-linux64.tar.gz
+    sudo ln -sf /opt/nvim-linux64/bin/nvim /usr/local/bin/nvim  # Make Neovim accessible via the nvim command
+    echo "Neovim installation completed."
+  fi
 }
 
 # Function to install fzf
